fs-4268 refactoring use of file after name change
