+++ tests/test_maintenance.py
@@ -7,10 +7,10 @@ from tests.conftest import fund_specific_claim_map
 @pytest.mark.mock_parameters(
     {
         "get_assessment_stats_path": [
-            "app.blueprints.assessments.models.fund_summary.get_assessments_stats",
+            "app.blueprints.assessments.models.round_summary.get_assessments_stats",
         ],
         "get_rounds_path": [
-            "app.blueprints.assessments.models.fund_summary.get_rounds",
+            "app.blueprints.assessments.models.round_summary.get_rounds",
         ],
         "fund_id": "test-fund",
         "round_id": "test-round",
